+++ trunk/hlfs/src/icache/icache_query.c
@@ -1,7 +1,7 @@
 #include "glib.h"
 #include "icache.h"
 
-iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint64_t iblock_no){
+iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint32_t iblock_no){
 	//HLOG_DEBUG("--Entering func %s", __func__);
 	int ret = 0;
 	if (icache_ctrl == NULL) {
@@ -10,7 +10,7 @@ iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint64_t iblock_no){
 		return NULL;
 	}
 	//HLOG_DEBUG("iblock_no %llu will be queried",iblock_no);
-    g_mutex_lock(icache_ctrl->icache_mutex);
+       g_mutex_lock(icache_ctrl->icache_mutex);
 	iblock_t * iblock = (iblock_t*)g_hash_table_lookup(icache_ctrl->iblock_map,&(iblock_no));
        g_mutex_unlock(icache_ctrl->icache_mutex);
        //if(_iblock!=NULL){
@@ -22,7 +22,7 @@ iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint64_t iblock_no){
 }
 
 
-int icache_query_iblock(ICACHE_CTRL *icache_ctrl, uint64_t iblock_no, char *iblock_buf){
+int icache_query_iblock(ICACHE_CTRL *icache_ctrl, uint32_t iblock_no, char *iblock_buf){
 	//HLOG_DEBUG("--Entering func %s", __func__);
 	int ret = 0;
 	icache_ctrl->total_read_count++;
