+++ trunk/hlfs/src/include/icache.h
@@ -7,7 +7,7 @@
 #include "comm_define.h"
 
 typedef struct {
-	uint64_t iblock_no;
+	uint32_t iblock_no;
 	char *   iblock;
 } iblock_t;
 
@@ -26,10 +26,10 @@ typedef struct icache_ctrl {
 	GTrashStack	*iblock_cache; 	//Stack used to store cache buffers
 	GQueue		*iblock_lru; 	//LRU queue of iblock
 	GHashTable	*iblock_map; 	//Hash Map
-	uint64_t 	       icache_size; 	//Number of cache buffers
-	uint64_t 	       iblock_size; 	//Size of each buffer
-	uint64_t		invalidate_trigger_level; 	// 
-	uint64_t		invalidate_once_size; 	//Number of dirty blocks have been written one time 
+	uint32_t 	       icache_size; 	//Number of cache buffers
+	uint32_t 	       iblock_size; 	//Size of each buffer
+	uint32_t		invalidate_trigger_level; 	// 
+	uint32_t		invalidate_once_size; 	//Number of dirty blocks have been written one time 
 	uint64_t 		total_write_count; 
 	uint64_t 		total_read_count;
     uint64_t 		icache_hit;
@@ -37,14 +37,14 @@ typedef struct icache_ctrl {
 
 ICACHE_CTRL *icache_new();
 int icache_init(ICACHE_CTRL *icache_ctrl,
-		uint64_t iblock_size,
-		uint64_t icache_size,
-		uint64_t invalidate_trigger_level,
-		uint64_t invalidate_once_size);
+		uint32_t iblock_size,
+		uint32_t icache_size,
+		uint32_t invalidate_trigger_level,
+		uint32_t invalidate_once_size);
 //int icache_insert_blocks(CACHE_CTRL *cache_ctrl, uint32_t start_block_no, uint32_t block_count,char *block_buf);
 int icache_insert_iblock(ICACHE_CTRL *icache_ctrl, uint32_t iblock_no, char *iblock_buf);
-iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint64_t iblock_no);
-int icache_query_iblock(ICACHE_CTRL *icache_ctrl, uint64_t iblock_no, char *iblock_buf);
+iblock_t * icache_query(ICACHE_CTRL *icache_ctrl,uint32_t iblock_no);
+int icache_query_iblock(ICACHE_CTRL *icache_ctrl, uint32_t iblock_no, char *iblock_buf);
 //gboolean  icache_iblock_exist(ICACHE_CTRL *icache_ctrl, uint64_t iblock_no);
 int icache_destroy(ICACHE_CTRL *icache_ctrl);
